Refactor smart projection factors around shared base
