+++ gtsam/slam/tests/testSmartProjectionPoseFactor.cpp
@@ -25,8 +25,10 @@
 #include <gtsam/nonlinear/LevenbergMarquardtOptimizer.h>
 #include <gtsam/slam/PoseTranslationPrior.h>
 #include <gtsam/slam/ProjectionFactor.h>
+#include <gtsam/slam/SmartProjectionFactor.h>
 
 #include <iostream>
+#include <type_traits>
 
 #include "smartFactorScenarios.h"
 
@@ -57,6 +59,25 @@ LevenbergMarquardtParams makeLMParams(bool verbose = false) {
 }
 
 }
+
+/* ************************************************************************* */
+namespace hierarchy {
+
+using Camera = PinholePose<Cal3_S2>;
+using ProjectionBase = SmartProjectionFactorBase<Camera>;
+using CameraFactor = SmartProjectionFactor<Camera>;
+using PoseFactor = SmartProjectionPoseFactor<Cal3_S2>;
+
+// Verifies that camera-variable and pose-variable factors share only the
+// projection-specific implementation base.
+TEST(SmartProjectionPoseFactor, CommonProjectionBase) {
+  EXPECT((std::is_base_of_v<ProjectionBase, CameraFactor>));
+  EXPECT((std::is_base_of_v<ProjectionBase, PoseFactor>));
+  EXPECT(!(std::is_base_of_v<CameraFactor, PoseFactor>));
+}
+
+}  // namespace hierarchy
+/* ************************************************************************* */
 
 /* ************************************************************************* */
 TEST( SmartProjectionPoseFactor, Constructor) {
